Moved passenger fragments to fragments/passenger
