+++ app/src/main/res/navigation/nav_graph_pass.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/home"
-        android:name="ua.com.cuteteam.cutetaxiproject.ui.passenger.fragments.PassengerFragment"
+        android:name="ua.com.cuteteam.cutetaxiproject.fragments.passenger.PassengerFragment"
         android:label="PassengerFragment" />
